+++ demisto_sdk/commands/content_graph/content_graph_builder.py
@@ -1,6 +1,7 @@
 import gc
 from typing import List, Optional
 
+import more_itertools
 import tqdm
 
 from demisto_sdk.commands.content_graph.common import Nodes, Relationships
@@ -46,16 +47,18 @@ class ContentGraphBuilder:
         self, packs_to_parse: Optional[List[str]] = None
     ) -> None:
 
-        content_dto: ContentDTO = self._create_content_dto(packs_to_parse)
-        self._collect_nodes_and_relationships_from_model(content_dto)
+        content_dtos: List[ContentDTO] = self._create_content_dtos(packs_to_parse)
+        for content_dto in content_dtos:
+            self._collect_nodes_and_relationships_from_model(content_dto)
 
-    def _create_content_dto(self, packs: Optional[List[str]]) -> ContentDTO:
+    def _create_content_dtos(self, packs: Optional[List[str]]) -> List[ContentDTO]:
         """Parses the repository, then creates and returns a repository model.
 
         Args:
             path (Path): The repository path.
             packs_to_parse (Optional[List[str]]): A list of packs to parse. If not provided, parses all packs.
         """
+        content_dtos = []
         repository_parser = RepositoryParser(self.content_graph.repo_path)
         packs_to_parse = tuple(repository_parser.iter_packs(packs))
         # parse the content packs with a progress bar
@@ -66,8 +69,13 @@ class ContentGraphBuilder:
             position=0,
             leave=True,
         ) as progress_bar:
-            repository_parser.parse(progress_bar=progress_bar)
-            return ContentDTO.from_orm(repository_parser)
+            for packs_batch in more_itertools.chunked(packs_to_parse, PACKS_PER_BATCH):
+                repository_parser.parse(packs_batch, progress_bar)
+                content_dtos.append(ContentDTO.from_orm(repository_parser))
+
+                repository_parser.clear()
+                gc.collect()
+        return content_dtos
 
     def _collect_nodes_and_relationships_from_model(
         self, content_dto: ContentDTO
@@ -75,12 +83,6 @@ class ContentGraphBuilder:
         for pack in content_dto.packs:
             self.nodes.update(pack.to_nodes())
             self.relationships.update(pack.relationships)
-
-        if (
-            conf_json := content_dto.conf_json
-        ):  # Repo could also not have it # TODO check
-            self.nodes.update(conf_json.to_node())
-            self.relationships.update(conf_json.relationships)
 
     def create_graph(self) -> None:
         self._parse_and_model_content()
