+++ demisto_sdk/commands/content_graph/parsers/json_content_item.py
@@ -24,7 +24,9 @@ class JSONContentItemParser(ContentItemParser):
     ) -> None:
         super().__init__(path, pack_marketplaces)
         self.path = self.get_path_with_suffix(".json")
-        self.git_sha = git_sha
+
+        self.json_data: Dict[str, Any] = self.get_json(git_sha=git_sha)
+        self.original_json_data: Dict[str, Any] = self.json_data
         if not isinstance(self.json_data, dict):
             raise InvalidContentItemException(
                 f"The content of {self.path} must be in a JSON dictionary format"
@@ -91,6 +93,5 @@ class JSONContentItemParser(ContentItemParser):
     def marketplaces(self) -> List[MarketplaceVersions]:
         return self.get_marketplaces(self.json_data)
 
-    @property
-    def json_data(self) -> Dict[str, Any]:
-        return get_json(str(self.path), git_sha=self.git_sha)
+    def get_json(self, git_sha: Optional[str]) -> Dict[str, Any]:
+        return get_json(str(self.path), git_sha=git_sha)
