+++ demisto_sdk/commands/content_graph/parsers/repository.py
@@ -8,7 +8,6 @@ from tqdm import tqdm
 from demisto_sdk.commands.common.constants import PACKS_FOLDER
 from demisto_sdk.commands.common.cpu_count import cpu_count
 from demisto_sdk.commands.common.logger import logger
-from demisto_sdk.commands.content_graph.parsers.conf_json import ConfJSONParser
 from demisto_sdk.commands.content_graph.parsers.pack import PackParser
 
 IGNORED_PACKS_FOR_PARSING = ["NonSupported"]
@@ -31,7 +30,6 @@ class RepositoryParser:
         """
         self.path: Path = path
         self.packs: List[PackParser] = []
-        self.conf_json = ConfJSONParser(self.path / "Tests" / "conf.json")
 
     def parse(
         self,
@@ -48,7 +46,6 @@ class RepositoryParser:
                     self.packs.append(pack)
                     if progress_bar:
                         progress_bar.update(1)
-
         except Exception:
             logger.error(traceback.format_exc())
             raise
