+++ demisto_sdk/commands/validate/default_config.toml
@@ -1,48 +1,75 @@
-[use_git]
+ignorable_errors = [
+    "BA101", "BA106", "BA108", "BA109", "BA110", "BA111", "BA112", "BA113", "BA116", "BA119", "BA124", "BA125",
+    "GF102",
+    "DS107", "DS108",
+    "IF100", "IF106", "IF113", "IF115", "IF116",
+    "IN107", "IN109", "IN110", "IN122", "IN123", "IN124", "IN126", "IN128", "IN135", "IN136", "IN139", "IN144", "IN145", "IN150", "IN153", "IN154", "IN161",
+    "MP106",
+    "PA113", "PA116", "PA124", "PA125", "PA127", "PA129",
+    "PB104", "PB105", "PB106", "PB107", "PB110", "PB111", "PB114", "PB115", "PB116", "PB118", "PB119", "PB121",
+    "RM100", "RM102", "RM104", "RM106", "RM108", "RM110", "RM112", "RM113",
+    "RP102", "RP104",
+    "SC100", "SC105", "SC106",
+    "IM111",
+    "RN112", "RN113", "RN114", "RN115", "RN116",
+    "MR108",
+    "PR101",
+    "LO107",
+    "DB100",
+    "GR103",
+    "SC105"
+]
 
+
+[use_git]
 select = [
-    "BA101", "BA105", "BA106", "BA116", "BA118",
-    "PA107", "PA108", "PA111", "PA115", "PA130",
+    "BA101", "BA105", "BA106", "BA116", "BA118", "BA126",
+    "PA100", "PA101", "PA102", "PA103", "PA104", "PA105", "PA107", "PA108", "PA109", "PA111", "PA113", "PA115", "PA117", "PA118", "PA119", "PA120",
+    "PA121", "PA123", "PA125", "PA127", "PA130",
     "BC100",
     "GR106",
-    "DO108",
-    "IN108",
-    "IN130",
+    "IN100", "IN102", "IN104", "IN106", "IN107", "IN108", "IN109", "IN110", "IN112", "IN113", "IN114", "IN115", "IN117", "IN118", "IN121", "IN122", "IN123", "IN124", "IN125", "IN126", "IN127", "IN130",
+    "IN131", "IN134", "IN135", "IN139", "IN141", "IN142", "IN145", "IN146", "IN149", "IN150", "IN151", "IN152", "IN153", "IN154", "IN156", "IN158", "IN159", "IN160",
+    "IN161", "IN162",
     "PB118",
-    "GR109",
-    "GR110",
+    "DO100", "DO101", "DO102", "DO103", "DO104, DO105, DO106",
+    "SC100", "SC105", "SC106", "SC109"
 ]
 warning = []
-ignorable_errors = [
-    "BA101", "BA116",
-    "PB118"
-]
+
 
 [validate_all]
-
 select = [
-    "BA101", "BA105", "BA106","BA116", "BA118",
+    "IN100", "IN102", "IN104", "IN106", "IN107", "IN108", "IN109", "IN110", "IN112", "IN113", "IN114", "IN121", "IN122", "IN123", "IN124", "IN125", "IN126", "IN127", "IN130",
+    "IN131", "IN134", "IN135", "IN139", "IN141", "IN142", "IN145", "IN146", "IN149", "IN150", "IN151", "IN152", "IN153", "IN154", "IN156", "IN158", "IN159", "IN160",
+    "IN161", "IN162",
+    "BA101", "BA105", "BA106","BA116", "BA118", "BA126",
     "GR106",
-    "DO108",
-    "IN108",
-    "IN130",
-    "PA107", "PA108", "PA111", "PA115", "PA130",
-    "GR109",
-    "GR110",
+    "PA100", "PA101", "PA102", "PA103", "PA104", "PA105", "PA107", "PA108", "PA109", "PA111", "PA113", "PA115", "PA117", "PA118", "PA119", "PA120",
+    "PA121", "PA123", "PA125", "PA127", "PA130",
+    "DO100", "DO101", "DO102", "DO103", "DO104",
+    "SC100", "SC105", "SC106", "SC109"
 ]
 warning = []
-ignorable_errors = ["BA101", "BA116"]
+
 
 [Custom_Categories]
 select = []
 warning = []
-ignorable_errors = []
 
 
 [support_level]
 
 [support_level.community]
-ignore = []
+ignore = [
+    "BC100", "BC101", "BC102", "BC103", "BC104", "BC105", "BC106", "BC107", "BC108", "BC109",
+    "CJ100", "CJ101", "CJ102", "CJ103", "CJ104", "CJ105",
+    "DS100", "DS101", "DS102", "DS103", "DS104",
+    "IN125", "IN126", "IN140"
+]
 
 [support_level.partner]
-ignore = []
+ignore = [
+    "CJ100", "CJ101", "CJ102", "CJ103", "CJ104", "CJ105",
+    "IN140"
+]
